Hotelproject/requirements.txt
# Project dependencies
djangorestframework
django-filter

# Optional: pin versions if needed, e.g.
# Django>=4.2